requirements.txt
psycopg2
pyTelegramBotAPI
pytest
python-dotenv
